Move users relations to concern User::Relations
