+++ app/models/user.rb
@@ -1,12 +1,6 @@
 class User < ApplicationRecord
   include Roles
-
-  has_one :api_key
-
-  has_many :comments
-  has_many :domains, foreign_key: :user_id
-  has_many :stamps, foreign_key: :user_id
-  has_many :votes
+  include Relations
 
   validates_presence_of %i[role username]
 
@@ -42,13 +36,5 @@ class User < ApplicationRecord
 
   def voted?(votable)
     votable.votes.where(user_id: id).exists?
-  end
-
-  def activities
-    PublicActivity::Activity.where(owner_id: self.id)
-  end
-
-  def domains
-    Domain.where(id: activities.where(key: 'domain.create').select(:trackable_id))
   end
 end
